requirements.txt
sherlock-project>=0.16.0
socialscan>=2.0.1
holehe>=1.61
